Project2/pub2/requirements.txt
kafka-python==2.0.2
yfinance==0.1.63
